scholorship_service_backend/requirements.txt
Django>=3.2,<4
psycopg2-binary>=2.8,<3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.1.0
dj-database-url==0.5.0
whitenoise==5.3.0
Pillow==8.3.2
django-cors-headers==4.3.0